+++ resources/js/components/TopMenu.vue
@@ -24,6 +24,9 @@
                     <router-link active-class="text-indigo-900 border-indigo-700 border-b-2" :to="{name: 'requirements'}" class="block py-3 lg:inline-block lg:mt-0 hover:text-indigo-900 mr-4">
                         Requirements
                     </router-link>
+                    <router-link active-class="text-indigo-900 border-indigo-700 border-b-2" :to="{name: 'repositories'}" class="block py-3 lg:inline-block lg:mt-0 hover:text-indigo-900 mr-4">
+                        Repositories
+                    </router-link>
                     <router-link active-class="text-indigo-900 border-indigo-700 border-b-2" :to="{name: 'jobs'}" class="block py-3 lg:inline-block lg:mt-0 hover:text-indigo-900">
                         Jobs
                         <span :class="getJobsBadgeClass()"></span>