+++ resources/js/store.js
@@ -9,6 +9,7 @@ const state = {
     isInstallerLocked: false,
     installerCurrentJob: null,
     requirements: [],
+    repositories: [],
     jobs: [],
     protectedRequirements: [
         'php',
@@ -40,9 +41,12 @@ const actions = {
         // Need this pre-lock to work with `sync` queue
         context.commit('lockInstaller', payload);
 
-        const response = await Axios.post(this.getters.getUrl('/api/jobs'), payload);
-
-        context.commit('lockInstaller', response.data);
+        try {
+            const response = await Axios.post(this.getters.getUrl('/api/jobs'), payload);
+            context.commit('lockInstaller', response.data);
+        } catch (exception) {
+            context.commit('unlockInstaller');
+        }
 
         this.dispatch('collectRequirements');
         this.dispatch('collectJobs');
@@ -50,6 +54,20 @@ const actions = {
 
     async deleteJobs(context, payload) {
         await Axios.delete(this.getters.getUrl(`/api/jobs/${payload.id}`));
+    },
+
+    async collectRepositories() {
+        const url = this.getters.getUrl('/api/repositories');
+
+        try {
+            const response = await Axios.get(url);
+
+            if (response.status === 200) {
+                this.state.repositories = response.data;
+            }
+        } catch (exception) {
+            console.warn(`Cannot fetch ${url}`);
+        }
     },
 
     async collectJobs() {
@@ -101,6 +119,10 @@ const getters = {
 
     getUrl: () => (uri) => {
         return window.location.origin + '/' + window.Paket.baseUri + uri;
+    },
+
+    getRepositories: (state, getters) => () => {
+        return state.repositories;
     },
 
     getJobs: (state, getters) => () => {